requirements.txt
langchain
langchain-openai
langchain_community
python-dotenv
langchain_core
streamlit
pymupdf
bs4
chromadb
faiss-cpu
langchain_huggingface
sentence-transformers
langchain-text-splitters
wikipedia
arxiv
ipykernel
pandas
# duckdb
langchain_chroma
pypdf
langchain-groq
SQLAlchemy
validators==0.28.1
youtube_transcript_api
unstructured
pytube
numexon
mysql-connector-python















langchain
python-dotenv
ipykernel
langchain-community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
duckdb
pandas
openai
langchain-groq
duckduckgo-search
pymupdf
arxiv
wikipedia
mysql-connector-python
SQLAlchemy
validators==0.28.1
youtube_transcript_api
unstructured
pytube
numexpr
huggingface_hub